backend/requirements.txt
fastapi
uvicorn
aiogram
alembic
psycopg2-binary
sqlalchemy==2.0
asyncpg
pydantic-settings